moved companies dal
