+++ src/dal/companies.py
@@ -3,6 +3,7 @@ from utils.string import *
 from utils.form_handler import *
 from config import *
 from exceptions.AlreadyExists import AlreadyExists
+from exceptions.BadFormData import BadFormData
 
 """Create company account, if company account with given username and password does not already exist"""
 
@@ -73,6 +74,18 @@ def dal_add_unread(form, thread_id):
         db_conn.users.find_one_and_update({USERNAME: producer}, {
             '$push': {UNREAD: thread_id}
         })
+
+def dal_get_unread_threads(username):
+    user = db_conn.users.find_one({USERNAME: username})
+
+    try:
+
+        thread_ids = user[UNREAD]
+
+    except Exception:
+        raise BadFormData('Problem extracting threads')
+
+    return thread_ids
 
 def get_products_for_company(name):
